refactor: move act method
